requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
chardet==5.2.0
charset-normalizer==3.4.0
cloudinary==1.41.0
cssselect2==0.7.0
Django==5.1.3
django-cloudinary-storage==0.3.0
django-environ==0.11.2
fonttools==4.55.2
idna==3.10
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.15.0
reportlab==4.2.5
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
tinycss2==1.4.0
tinyhtml5==2.0.0
urllib3==2.2.3
weasyprint==63.0
webencodings==0.5.1
zopfli==0.2.3.post1
